requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
config==0.5.1
decouple==0.0.7
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pep8==1.7.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
uritemplate==4.2.0
